+++ sdks/python/apache_beam/dataframe/io_test.py
@@ -455,9 +455,9 @@ class ReadGbqTransformTests(unittest.TestCase):
   def test_ReadGbq_unsupported_param(self):
     with self.assertRaisesRegex(
         ValueError,
-        "Unsupported parameter entered in read_gbq. "
-        "Please enter only supported parameters 'table', "
-        "'dataset', 'project_id', 'use_bqstorage_api'."):
+        "Unsupported parameter entered in read_gbq. Please enter only "
+        "supported parameters 'table', 'dataset', "
+        "'project_id', 'use_bqstorage_api'."):
       p = beam.Pipeline()
       _ = p | beam.dataframe.io.read_gbq(
           table="table", use_bqstorage_api=False, reauth="true_config")
